Limit zooming to avoid problems with tick calculation
Reject scaling if either the tick interval is less that epsilon or if either the high or
low coordinate divided by the interval is NaN or Inf. This fixes issue #5.

The y-axis position depends on the length of the y-tick labels.  Set the left end of the
x-axis after setting the position of the y-axis.