Bug fixes and documentation updates.
Parameterized panning and zooming.

Fix issue 7 Clicking near the origin changes the range. Clip the pointer position to the
grid. Redraw the entire plot when clicking near the origin because that covers the whole
range with a zoom box.

Add a unit of width when redrawing stroked rectangles to avoid fringes.

Change from Ctrl to Alt for constraining to horizontal mouse wheel scrolling. This allows
Ctrl to be used for coarse control for consistency.

Documented key sequences. Added layout diagram.