Added choice of color palette.
ggplot2, Brewer set 1 and set 2, and Okabe-Ito paletts are available. ggplot2 is the
default. Can be set with the -p or --palette option. The R function pview() accepts a
palette option. The option strings are 'ggplot2', 'set1', 'set2', and 'okabe-ito'.

Moved color code to color.cc and .hh.

Plot points as circles (round endcaps). But if there are > 5000 points, use squares (square
endcaps) which are quicker to draw.